+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - updated project selection logic when closing or adding projects
 - replaced react-table library with custom implementation to avoid incompatibility with React Compiler
 - added Delegated task state when handing off the task to subtask
+- hide task state actions when no messages are present
 
 ## [0.60.0]